Added additional test for EventList and then refactored code for better organization by introducing mountApp and then added wait-for-expect
