+++ package.json
@@ -22,9 +22,6 @@
   },
   "devDependencies": {
     "@babel/core": "^7.6.0",
-    "babel-preset-env": "^1.7.0",
-    "sass": "^1.22.12",
-    "vue-jest": "^3.0.5",
     "@vue/cli-plugin-babel": "^3.11.0",
     "@vue/cli-plugin-e2e-cypress": "^3.11.0",
     "@vue/cli-plugin-eslint": "^3.11.0",
@@ -35,12 +32,16 @@
     "babel-core": "7.0.0-bridge.0",
     "babel-eslint": "^10.0.1",
     "babel-jest": "^23.6.0",
+    "babel-preset-env": "^1.7.0",
     "eslint": "^5.16.0",
     "eslint-plugin-prettier": "^3.1.0",
     "eslint-plugin-vue": "^5.0.0",
     "node-sass": "^4.9.0",
     "prettier": "^1.18.2",
+    "sass": "^1.22.12",
     "sass-loader": "^7.1.0",
-    "vue-template-compiler": "^2.6.10"
+    "vue-jest": "^3.0.5",
+    "vue-template-compiler": "^2.6.10",
+    "wait-for-expect": "^2.0.0"
   }
 }
